# docs/swagger.yaml
definitions:
...
  models.Role:
    properties:
      descripcion:
        type: string
      nombre_rol:
        type: string
    type: object
...
  models.User:
    properties:
      contraseña:
        type: string
      id_rol:
        description: Llave foránea que referencia la tabla roles
        type: integer
      nombre_usuario:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Carga la información del rol relacionado
    type: object
...
  request.CreateRoleDTO:
    properties:
      descripcion:
        type: string
      nombre_rol:
        type: string
    required:
    - descripcion
    - nombre_rol
    type: object
...
  request.CreateUserDTO:
    properties:
      contraseña:
        minLength: 8
        type: string
      id_rol:
        description: ID del rol asociado, requerido
        type: integer
      nombre_usuario:
        type: string
    required:
    - contraseña
    - id_rol
    - nombre_usuario
    type: object
info:
  contact: {}
paths:
...
  /roles:
    get:
      description: Retrieve a list of all roles in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
      summary: List all roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Adds a new role to the system
      parameters:
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/request.CreateRoleDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new role
      tags:
      - roles
  /roles/{id}:
    delete:
      description: Remove a role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a role
      tags:
      - roles
    get:
      description: Retrieve a single role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update the information of an existing role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/request.CreateRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a role
      tags:
      - roles
...
  /users:
    get:
      description: Retrieve a list of all users in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Adds a new user to the system
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Remove a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      description: Retrieve a single user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the information of an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user
      tags:
      - users
swagger: "2.0"