# internal/API/dto/request/user_request_dto.go
package request

type CreateUserDTO struct {
	IDRole   uint   `json:"id_rol" binding:"required"` // ID del rol asociado, requerido
	UserName string `json:"nombre_usuario" binding:"required"`
	Password string `json:"contraseña" binding:"required,min=8"`
}